0.1.3.Nightly.1
Added NetworkObject.IsNetworked. When false on the prefab or inspector the object will not initialize as a networked object. When an object is passed in using Spawn it automatically becomes a networked object. This can be useful if you have an object that contains networked scripts but may also serve as a non-networked object.
Added ObjectId QOL to NetworkBehaviour.
Added TimeManager QOL to NetworkBehaviour and NetworkObject.
Added additional TicksToTime methods in TimeManager.
Begun prediction work; some files exist but are incomplete.
Added upgrade tool for Mirror to Fish-Networking. See menu Fish-Networking.